Offset doesn't need such a large int type

+++ Sources/Meter/CallStackTree.swift
@@ -43,14 +43,14 @@ public struct Binary: Codable, Hashable {
 
 public struct Frame: Codable, Hashable {
     public var binaryUUID: UUID?
-    public var offsetIntoBinaryTextSegment: UInt64?
+    public var offsetIntoBinaryTextSegment: Int?
     public var sampleCount: Int?
     public var binaryName: String?
     public var address: UInt64
     public var subFrames: [Frame]?
     public var symbolInfo: [SymbolInfo]?
 
-    public init(binaryUUID: UUID? = nil, offsetIntoBinaryTextSegment: UInt64? = nil, sampleCount: Int? = nil, binaryName: String? = nil, address: UInt64, subFrames: [Frame]?, symbolInfo: [SymbolInfo]? = nil) {
+    public init(binaryUUID: UUID? = nil, offsetIntoBinaryTextSegment: Int? = nil, sampleCount: Int? = nil, binaryName: String? = nil, address: UInt64, subFrames: [Frame]?, symbolInfo: [SymbolInfo]? = nil) {
         self.binaryUUID = binaryUUID
         self.offsetIntoBinaryTextSegment = offsetIntoBinaryTextSegment
         self.sampleCount = sampleCount
@@ -82,12 +82,12 @@ public struct Frame: Codable, Hashable {
             return nil
         }
 
-		let loadAddress = withOffsetAsLoadAddress ? offset : address - offset
-        let size = address - loadAddress + 1
+		let loadAddress = withOffsetAsLoadAddress ? offset : Int(address) - offset
+        let size = Int(address) - loadAddress + 1
 
         return Binary(uuid: uuid,
-                      loadAddress: Int(loadAddress),
-                      approximateSize: Int(size),
+                      loadAddress: loadAddress,
+                      approximateSize: size,
                       name: binaryName)
     }
